[refactor] refactor package structure
- refactor package structure (to use domain package structrue)
