+++ back/demo/src/main/java/OOTD/demo/auth/controller/AuthController.java
@@ -1,14 +1,14 @@
-package OOTD.demo.api.controller;
+package OOTD.demo.auth.controller;
 
-import OOTD.demo.api.dto.auth.CheckNameDto;
-import OOTD.demo.api.dto.auth.CreateUserDto;
-import OOTD.demo.api.dto.auth.LoginDto;
-import OOTD.demo.api.dto.auth.CheckPasswordDto;
+import OOTD.demo.auth.dto.CheckNameDto;
+import OOTD.demo.auth.dto.CreateUserDto;
+import OOTD.demo.auth.dto.LoginDto;
+import OOTD.demo.auth.dto.CheckPasswordDto;
 import OOTD.demo.config.security.JwtTokenProvider;
-import OOTD.demo.domain.Message;
-import OOTD.demo.domain.user.User;
-import OOTD.demo.service.user.AuthService;
-import OOTD.demo.service.user.UserService;
+import OOTD.demo.common.Message;
+import OOTD.demo.user.User;
+import OOTD.demo.auth.service.AuthService;
+import OOTD.demo.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
