+++ back/demo/src/main/java/OOTD/demo/auth/dto/LoginDto.java
@@ -1,4 +1,4 @@
-package OOTD.demo.api.dto.auth;
+package OOTD.demo.auth.dto;
 
 
 import lombok.Getter;
@@ -6,6 +6,7 @@ import lombok.Setter;
 
 @Getter
 @Setter
-public class CheckPasswordDto {
+public class LoginDto {
+    private String email;
     private String password;
 }
