+++ back/demo/src/main/java/OOTD/demo/config/WebSecurityConfig.java
@@ -1,6 +1,5 @@
 package OOTD.demo.config;
 
-import OOTD.demo.service.user.AuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
