+++ back/demo/src/main/java/OOTD/demo/diary/Diary.java
@@ -2,7 +2,7 @@ package OOTD.demo.diary;
 
 import OOTD.demo.diary.dto.PostDiaryReqDTO;
 import OOTD.demo.diary.dto.UpdateDiaryReqDTO;
-import OOTD.demo.domain.user.User;
+import OOTD.demo.user.User;
 import lombok.Getter;
 import javax.persistence.*;
 import java.time.LocalDateTime;
