+++ back/demo/src/main/java/OOTD/demo/diary/repository/DiaryRepository.java
@@ -1,6 +1,6 @@
 package OOTD.demo.diary.repository;
 
-import OOTD.demo.domain.user.User;
+import OOTD.demo.user.User;
 import OOTD.demo.diary.Diary;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
