+++ back/demo/src/main/java/OOTD/demo/user/dto/ReadUserDto.java
@@ -1,4 +1,4 @@
-package OOTD.demo.api.dto.user;
+package OOTD.demo.user.dto;
 
 
 import lombok.Getter;
